实现knock gate功能

+++ main/src/player.rs
@@ -60,8 +60,8 @@ pub struct Player {
     pub accuracy: u32,  //准确度  
     pub avoidance: u32, //规避力
     pub timer_id: usize, //定时器ID
-    pub pending: usize,  //暂停
-    pub knocked: usize,  //是否已经敲门 0:否  1:是
+    pub pending: u32,  //暂停
+    pub knocked: u32,  //是否已经敲门 0:否  1:是
 }
